restructuring and adding basic testing skel
