+++ test/fixtures/gitlab.rb
