+++ test/integration/helpers/serverspec/spec_helper.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require 'serverspec'
+
+set :backend, :exec
